+++ vacefron/__init__.py
@@ -1,3 +1,4 @@
 from .client import Client
 from .errors import BadRequest, HTTPException, InternalServerError, NotFound
-from .image import Image
+
+__version__ = '1.1.0'