+++ vacefron/client.py
@@ -1,14 +1,13 @@
 from asyncio import get_event_loop
-from re import search
 from typing import Any, Tuple
 
 from aiohttp import ClientSession
 
+from .classes import Image, RankCard
 from .errors import BadRequest, HTTPException, InternalServerError, NotFound
-from .image import Image
 
 
-def _parse_text(text: str) -> str:
+def _replace_characters(text: str) -> str:
     replacements = {
         " ": "%20",
         "!": "%21",
@@ -33,18 +32,18 @@ def _parse_text(text: str) -> str:
         "^": "%CB%86",
         "_": "%5F",
         "©": "%C2%A9"
-    }
+        }
     return text.translate(str.maketrans(replacements))
 
 
 class Client:
 
-    def __init__(self, session: ClientSession = None) -> None:
+    def __init__(self, session: object = None) -> None:
         self.session = ClientSession(loop = get_event_loop()) or session
         self._api_url = "https://vacefron.nl/api"
 
-    async def _check_url(self, url: str):  # return_response=False):
-        response = await self.session.get(url)
+    async def _check_url(self, url: str):
+        response = await self.session.get(str(url))
         if response.status == 200:
             return url
         elif response.status == 400:
@@ -70,13 +69,22 @@ class Client:
 
     # Image
 
+    async def distracted_bf(self, boyfriend: str, woman: str, girlfriend: str) -> Image:
+        url = f"{self._api_url}/distractedbf" \
+              f"?boyfriend={boyfriend}" \
+              f"&woman={woman}" \
+              f"&girlfriend={girlfriend}"
+
+        url = await self._check_url(url)
+        return Image(url, self.session)
+
     async def car_reverse(self, text: str) -> Image:
-        text = _parse_text(str(text))
+        text = _replace_characters(str(text))
         url = await self._check_url(f"{self._api_url}/carreverse?text={text}")
         return Image(url, self.session)
 
     async def change_my_mind(self, text: str) -> Image:
-        text = _parse_text(str(text))
+        text = _replace_characters(str(text))
         url = await self._check_url(f"{self._api_url}/changemymind?text={text}")
         return Image(url, self.session)
 
@@ -94,9 +102,10 @@ class Client:
 
     async def i_can_milk_you(self, user: str, user2: str = None) -> Image:
         url = f"{self._api_url}/icanmilkyou?user1={user}"
-        url = await self._check_url(url)
         if user2 is not None:
             url += f"&user2={user2}"
+
+        url = await self._check_url(url)
         return Image(url, self.session)
 
     async def heaven(self, user: str) -> Image:
@@ -104,6 +113,8 @@ class Client:
         return Image(url, self.session)
 
     async def npc(self, text: str, text2: str) -> Image:
+        text = _replace_characters(str(text))
+        text2 = _replace_characters(str(text2))
         url = await self._check_url(f"{self._api_url}/npc?text1={text}&text2={text2}")
         return Image(url, self.session)
 
@@ -116,7 +127,7 @@ class Client:
         return Image(url, self.session)
 
     async def water(self, text: str) -> Image:
-        text = _parse_text(str(text))
+        text = _replace_characters(str(text))
         url = await self._check_url(f"{self._api_url}/water?text={text}")
         return Image(url, self.session)
 
@@ -124,25 +135,21 @@ class Client:
         url = await self._check_url(f"{self._api_url}/wide?image={image}")
         return Image(url, self.session)
 
-    async def rank_card(self, username: str, avatar: str, level: int, rank: int, current_xp: int,
-                        next_level_xp: int, previous_level_xp: int, custom_background: str = None,
-                        xp_color: str = None, is_boosting: bool = False) -> Image:
-        username = _parse_text(str(username))
-        url = f"?username={username}&avatar={avatar}&level={level}&rank={rank}&currentxp={current_xp}" \
-              f"&nextlevelxp={next_level_xp}&previouslevelxp={previous_level_xp}"
+    async def rank_card(
+            self, username: str, avatar: str, level: int, rank: int, current_xp: int,
+            next_level_xp: int, previous_level_xp: int, custom_background: str = None,
+            xp_color: str = None, is_boosting: bool = False) -> Image:
 
-        if custom_background is not None:
-            url += f"&custombg={custom_background}"
-        if xp_color is not None:
-            if not search(r'^(?:[0-9a-fA-F]{3}){1,2}$', str(xp_color)):
-                raise BadRequest("Invalid HEX value. You're only allowed to enter HEX (0-9 & A-F)")
-            else:
-                url += f"&xpcolor={xp_color}"
-        if is_boosting:
-            url += f"&isboosting=true"
+        username = _replace_characters(str(username))
+        xp_color = str(xp_color).replace("#", "") if xp_color else None
 
-        actual_url = await self._check_url(f"{self._api_url}/rankcard{url}")
-        return Image(actual_url, self.session)
+        card = RankCard(
+                username, avatar, level, rank, current_xp,
+                next_level_xp, previous_level_xp, custom_background, xp_color, is_boosting
+                )
+
+        await self._check_url(str(card))
+        return Image(str(card), self.session)
 
     async def close(self) -> None:
         if not self.session.closed: