+++ changelog.md
@@ -1,6 +1,11 @@
 # VACEfron.py | Changelog
 See here what changed or broke each version.
 
+## v1.1.0
+- Added support for new endpoint `.distracted_bf()`. See the [Docs](docs.md#await-vac_apidistracted_bfboyfriend-woman-girlfriend) for it.
+- Renamed `image.py` to `classes.py`
+- Added a RankCard class, nothing interesting, keep using `.rank_card()`.
+
 ## v1.0.2
 - docs.md and README.md improvements. New version to update it on PyPi.